AddNewTask page tests created
