+++ TeamProjectFluxday/TeamProjectFluxday/Pages/MyTasks/AddNewTask/AddNewTaskPageMap.cs
@@ -1,0 +1,32 @@
+﻿using OpenQA.Selenium;
+using TeamProjectFluxday.Core;
+
+namespace TeamProjectFluxday.Pages.MyTasks.AddNewTask
+{
+    public class AddNewTaskPageMap : BasePageMap
+    {
+        public IWebElement TitleInput
+        {
+            get
+            {
+                return GetElement(By.CssSelector("#task_name"));
+            }
+        }
+
+        public IWebElement DescriptionInput
+        {
+            get
+            {
+                return GetElement(By.CssSelector("#task_description"));
+            }
+        }
+
+        public IWebElement CreateTaskButton
+        {
+            get
+            {
+                return GetElement(By.CssSelector(".button"));
+            }
+        }
+    }
+}
