+++ TeamProjectFluxday/TeamProjectFluxday/Pages/MyTasks/AddNewTask/AddNewTaskPageValidator.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using TeamProjectFluxday.Core;
+using TeamProjectFluxday.Data;
+
+namespace TeamProjectFluxday.Pages.MyTasks.AddNewTask
+{
+    public class AddNewTaskPageValidator : BasePageValidator<AddNewTaskPageMap>
+    {
+      
+    }
+}
