+++ TeamProjectFluxday/TeamProjectFluxday/Pages/MyTasks/MyTasksPage.cs
@@ -1,6 +1,5 @@
 ﻿using System.Threading;
 using TeamProjectFluxday.Core;
-using TeamProjectFluxday.Data;
 
 namespace TeamProjectFluxday.Pages
 {
@@ -8,17 +7,6 @@ namespace TeamProjectFluxday.Pages
     {
         public MyTasksPage(string url = "https://app.fluxday.io/tasks#pane2") : base(url)
         {
-        }
-
-        internal void CreateNewTask()
-        {
-            var newTask = TestData.Task;
-
-            Map.TitleInput.SendKeys(newTask.Title);
-
-            Map.DescriptionInput.SendKeys(newTask.Description);
-
-            Map.CreateTaskButton.Click();
         }
 
         internal void DeleteNewTask()
