+++ TeamProjectFluxday/TeamProjectFluxday/Pages/MyTasks/MyTasksPageMap.cs
@@ -5,36 +5,12 @@ namespace TeamProjectFluxday.Pages
 {
     public class MyTasksPageMap : BasePageMap
     {
-        public IWebElement TitleInput
-        {
-            get
-            {
-                return GetElement(By.CssSelector("#task_name"));
-            }
-        }
-
-        public IWebElement DescriptionInput
-        {
-            get
-            {
-                return GetElement(By.CssSelector("#task_description"));
-            }
-        }
-
-        public IWebElement CreateTaskButton
-        {
-            get
-            {
-                return GetElement(By.CssSelector(".button"));
-            }
-        }
-
         public IWebElement TaskTitle
         {
             get
             {
                 return GetElement(By.XPath(@"//*[@id='paginator']/a/div/div[1]"));
-            }                        
+            }
         }
 
         public IWebElement NewTask
