+++ TeamProjectFluxday/TeamProjectFluxday/Tests/AddNewTaskPageTests.cs
@@ -1,0 +1,36 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using TeamProjectFluxday.Core;
+using TeamProjectFluxday.Data;
+using TeamProjectFluxday.Pages;
+using TeamProjectFluxday.Utils;
+
+namespace TeamProjectFluxday.Tests
+{
+    [TestClass]
+    public class AddNewTaskPageTests : BaseTest
+    {
+        DashboardPage dashboardPage;
+
+        [TestInitialize]
+        public void TestSetup()
+        {
+            Driver.StartBrowser();
+            dashboardPage = LoginProvider.Login(TestData.TeamLeadUser);
+        }
+
+        [TestCategory("AddNewTaskPageTests")]
+        [Owner("Yusuf Machan")]
+        [TestMethod]
+        public void Test001CreateNewTask()
+        {
+            var addNewTaskPage = dashboardPage.NavigationPanel.NavigateToAddNewTaskPage();
+            
+            addNewTaskPage.CreateNewTask();
+
+            var myTasksPage = dashboardPage.NavigationPanel.NavigateToMyTasksPage();
+
+            myTasksPage.Validate().TaskExists();
+            myTasksPage.DeleteNewTask();
+        }
+    }
+}
